+++ src/app/modules/utilities.js
@@ -1,7 +1,5 @@
 /* eslint-disable linebreak-style */
 export const ul = document.querySelector('.listContainer');
-export const overlay = document.querySelector('.overlay');
-export const popUp = document.querySelector('.pop-up');
 const baseURL = 'https://api.punkapi.com/v2/';
 const likesBaseURL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/';
 export const appID = 'cuSc6vNb8MIaUDaK4qjK';
@@ -20,9 +18,4 @@ export default async function apiCall(urlExtension = '', method = '', body, like
 export const error = (message = '') => {
   const error = `An error has ocurred ${message}`;
   console.log(error);
-};
-
-export const closePopUpHandler = (closePopUpBtn) => {
-  overlay.classList.add('hidden');
-  closePopUpBtn.removeEventListener('click', closePopUpHandler);
 };
