+++ src/app/scss/base/_reset.scss
@@ -28,4 +28,3 @@ ul {
   text-decoration: line-through;
   color: #bbb;
 }
-
