+++ src/app/scss/components/_index.scss
