+++ src/app/scss/components/_pop.scss
@@ -1,4 +1,4 @@
-.overlay{
+.overlay {
   position: absolute;
   top: 0;
   bottom: 0;
@@ -15,6 +15,7 @@
     width: 30%;
     position: relative;
     transition: all 5s ease-in-out;
+
     .close-pop-up {
       position: absolute;
       top: 20px;
@@ -25,6 +26,7 @@
       text-decoration: none;
       color: #333;
     }
+
     .pop-content {
       max-height: 30%;
       overflow: auto;
